backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.1
pymongo==4.4.1
python-dotenv==1.0.0
papermill==2.4.0
pytest==7.4.0
jupyter==1.0.0
notebook==6.5.4
ipykernel==6.25.1 